functions/pay_with_stripe/requirements.txt
google-cloud-firestore==0.30.0
google-cloud-pubsub==0.39.0
stripe==2.15.0